- Added reminder date pickers - Implemented datepicker style
