+++ client/package.json
@@ -11,10 +11,10 @@
         "final-form": "^4.19.1",
         "graphql": "14.3.1",
         "moment": "2.24.0",
-        "momentjs": "^2.0.0",
         "node-sass": "^4.13.1",
         "normalize.css": "^8.0.1",
         "react": "16.8.6",
+        "react-datepicker": "^2.14.1",
         "react-dom": "16.8.6",
         "react-final-form-hooks": "^2.0.2",
         "react-lottie": "^1.2.3",
