+++ client/src/assets/scss/base/_reset.scss
@@ -26,3 +26,5 @@ input, button {
 *:focus {
   outline: none;
 }
+
+
