+++ client/src/assets/scss/blocks/_form.scss
@@ -90,5 +90,3 @@
   color: $color-error;
   font-size: $font-size-sm;
 }
-
-
