+++ client/src/assets/scss/blocks/_react-datepicker.scss
@@ -1,0 +1,45 @@
+.react-datepicker {
+  border: none;
+  box-shadow: $box-shadow-strong;
+  color: $text-color-primary;
+}
+
+.react-datepicker__triangle {
+  display: none;
+}
+
+.react-datepicker__header {
+  background-color: $background-color-grey-light;
+  border: none;
+  padding: $spacing-def $spacing-sm 0 $spacing-xs;
+}
+
+.react-datepicker__day--selected {
+  background-color: $background-color-blue;
+  border-radius: $border-radius-def;
+}
+
+.react-datepicker__navigation--previous {
+  border-right-color: $color-blue;
+  left: 12px;
+  top: 17px;
+}
+
+.react-datepicker__navigation--next {
+  border-left-color: $color-blue;
+  right: 12px;
+  top: 17px;
+}
+
+.react-datepicker__day-names {
+  padding: $spacing-sm 0;
+}
+
+.react-datepicker__month {
+  margin: 0;
+  padding: $spacing-sm 0 $spacing-xs 0;
+}
+
+.react-datepicker-wrapper {
+  width: 100%;
+}
