+++ client/src/assets/scss/styles.scss
@@ -30,6 +30,7 @@
 @import "blocks/dashboard";
 @import "blocks/reminders";
 @import "blocks/todo-area";
+@import "blocks/react-datepicker";
 
 // ** REUSABLE BLOCKS **
 @import "blocks/reusable/bottom-info";
