+++ client/src/assets/scss/variables/_box-shadows.scss
@@ -1,2 +1,3 @@
+$box-shadow-strong: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
 $box-shadow-default: 0 2px 10px 0 $box-shadow-color-default;
 $box-shadow-default-focused: 0 0 0 4px $box-shadow-color-focused-large;
