+++ client/src/modules/Dashboard/Reminders/ReminderDialog/ReminderDialog.types.ts
@@ -1,0 +1,4 @@
+export type ReminderDialogProps = {
+    toggleDialog: () => void,
+    isDialogOpen: boolean,
+}
