+++ client/src/modules/Dashboard/Reminders/ReminderDialog/index.ts
@@ -1,0 +1,2 @@
+export * from './ReminderDialog'
+export * from './ReminderDialog.types'
