+++ client/tslint.json
@@ -12,6 +12,7 @@
         ],
         "interface-over-type-literal": false,
         "jsx-boolean-value": false,
+        "jsx-no-lambda": false,
         "jsx-no-multiline-js": false,
         "max-line-length": [
             true,
